Edit: Separate moredropdown component + css module file
