+++ src/components/moredropdown/MoreDropdown.js
@@ -2,7 +2,7 @@ import React from "react";
 import Dropdown from "react-bootstrap/Dropdown";
 import { useHistory } from "react-router";
 
-import styles from "../styles/MoreDropdown.module.css";
+import styles from "./MoreDropdown.module.css";
 
 /**
  * Code provided by Code Institute's "Moments" walkthrough.
